requirements.txt
pre-commit
homeassistant
AWSIoTPythonSDK~=1.5.2
aiohttp
cryptography
voluptuous
paho-mqtt~=1.6.1
requests~=2.28.1
